massive refactor
Lots of vars to hold the fns to call to render or act on entries.  I'm
not sure this is the right direction or approach, but it is worth
trying out.

Also lots of hooks for various behaviors instead of baking them in.

I'm also beginning to remove the homegrown html stripping functions.