Removed the call to get the unread count for the header
There was a race condition when marking a bunch of items as read and
it would often result in (xx/yy) showing up in the header with bogus
numbers. Thinking about it, since entries load in batches until all
are fetched, fetching the unread count separately probably isn't
necessary in the common case.